requirements.txt
Flask>=2.2,<3.0
Flask-SQLAlchemy>=3.0,<4.0
Flask-Login>=0.6,<1.0
Flask-Babel>=3.0,<4.0
Flask-CORS>=4.0,<5.0
Flask-JWT-Extended>=4.5,<5.0
Flask-Migrate>=4.0,<5.0
Werkzeug>=2.2,<3.0
requests>=2.28,<3.0
python-dotenv>=1.0,<2.0
email-validator>=2.0,<3.0
pytest>=7.4,<8.0
pytest-cov>=4.1,<5.0
Pillow>=10.0,<11.0
gunicorn>=21.2,<22.0
